remove warnings for compute options
